Fixed "Shift-release" issue
Note: I initially extracted `pure-functions.js` with the intention of integrating a testing framework and doing some TDD, but it's so much easier to reason about pure functions that I ended up fixing the problem too easily to be bothered with setting up a test harness. #toolazyiguess ¯\_(ツ)_/¯

---

Steps to Repro (in prior commits):

- Press and hold `Shift`
- Press and hold 'a'
  → heldKeys contains 'Shift' and 'A'
- Release `a`
- Release `Shift`
  → Expected: heldKeys is now empty
  → Actual heldKeys is now empty ✅

- Press and hold `Shift`
- Press and hold 'a'
  → heldKeys contains 'Shift' and 'A'
- Release `Shift`
- Release `a`
  → Expected: heldKeys is now empty
  → Actual: heldKeys still contains 'A' ❌

Diagnosis: When Shift is released first, the "key" reported by the browser goes from 'A' to 'a'. Since the set doesn't have 'a' in it, trying to remove it ends up in a no-op, leaving the 'A' stranded, still in the set. Of course, this commit changes `heldKeys` to an Array, so you'll have to look at prior commits to see what I mean.